# backend/src/translation-api-key/dto/update-translation-api-key.dto.ts
import { CreateTranslationApiKeyDto } from "./create-translation-api-key.dto";
import { PartialType } from '@nestjs/swagger';

export class UpdateTranslationApiKeyDto extends PartialType(CreateTranslationApiKeyDto) {
 
}