# backend/swagger.json
{"openapi":"3.0.0","paths":{"/":{"get":{"operationId":"AppController_getHello","parameters":[],"responses":{"200":{"description":""}}}},"/upload":{"post":{"operationId":"AppController_uploadFile","summary":"Upload an image","parameters":[],"requestBody":{"required":true,"content":{"multipart/form-data":{"schema":{"type":"object","properties":{"file":{"type":"string","format":"binary"}}}}}},"responses":{"200":{"description":"response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResponseDto"}}}},"201":{"description":""}},"security":[{"bearer":[]}]}},"/files/{filepath}":{"get":{"operationId":"AppController_seeUploadedFile","summary":"Visualize uploaded file","parameters":[{"name":"filepath","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"security":[{"bearer":[]}]}},"/products":{"post":{"operationId":"ProductController_create","parameters":[],"responses":{"201":{"description":""}}},"get":{"operationId":"ProductController_findAll","parameters":[{"name":"page","required":true,"in":"query","schema":{"type":"string"}},{"name":"pageSize","required":true,"in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":""}}}},"/products/{id}":{"get":{"operationId":"ProductController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}}},"patch":{"operationId":"ProductController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}}},"delete":{"operationId":"ProductController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}}}},"/collections":{"post":{"operationId":"CollectionController_create","parameters":[],"responses":{"201":{"description":""}},"security":[{"bearer":[]}]},"get":{"operationId":"CollectionController_findAll","parameters":[],"responses":{"200":{"description":""}},"security":[{"bearer":[]}]}},"/collections/list":{"post":{"operationId":"CollectionController_getList","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RequestCollectionListDto"}}}},"responses":{"201":{"description":""}},"security":[{"bearer":[]}]}},"/collections/{id}":{"get":{"operationId":"CollectionController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"security":[{"bearer":[]}]},"patch":{"operationId":"CollectionController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"security":[{"bearer":[]}]},"delete":{"operationId":"CollectionController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"security":[{"bearer":[]}]}},"/article":{"get":{"operationId":"ArticleController_create","parameters":[],"responses":{"200":{"description":""}}}},"/mail":{"get":{"operationId":"MailController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/EmailDTO"}}}},"responses":{"200":{"description":""}},"tags":["mail"],"security":[{"bearer":[]}]}},"/user":{"post":{"operationId":"UserController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserDTO"}}}},"responses":{"200":{"description":"users response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResponseDto"}}}}},"tags":["user"],"security":[{"bearer":[]}]},"get":{"operationId":"UserController_findAll","parameters":[],"responses":{"200":{"description":"users response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResponseDto"}}}}},"tags":["user"],"security":[{"bearer":[]}]}},"/user/{id}":{"get":{"operationId":"UserController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"users response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResponseDto"}}}}},"tags":["user"],"security":[{"bearer":[]}]},"patch":{"operationId":"UserController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserDTO"}}}},"responses":{"200":{"description":"users response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResponseDto"}}}}},"tags":["user"],"security":[{"bearer":[]}]},"delete":{"operationId":"UserController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"users response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResponseDto"}}}}},"tags":["user"],"security":[{"bearer":[]}]}},"/user/verfieUser/{id}":{"patch":{"operationId":"UserController_verifieUser","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"users response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResponseDto"}}}}},"tags":["user"],"security":[{"bearer":[]}]}},"/auth/login":{"post":{"operationId":"AuthController_login","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginDto"}}}},"responses":{"200":{"description":"login response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuthResponseDto"}}}}},"tags":["auth"]}},"/auth/logout":{"post":{"operationId":"AuthController_logout","parameters":[],"responses":{"201":{"description":""}},"tags":["auth"]}},"/auth/register":{"post":{"operationId":"AuthController_register","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserDTO"}}}},"responses":{"200":{"description":"login response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuthResponseDto"}}}}},"tags":["auth"]}},"/auth/refresh-token":{"post":{"operationId":"AuthController_refreshToken","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RefreshTokenDto"}}}},"responses":{"200":{"description":"login response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuthResponseDto"}}}}},"tags":["auth"],"security":[{"bearer":[]}]}},"/auth/request-reset-password-email":{"post":{"operationId":"AuthController_requestPasswordReset","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResetPasswordReqDto"}}}},"responses":{"200":{"description":"Request reset password","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResetPasswordDto"}}}}},"tags":["auth"]}},"/auth/reset-password":{"post":{"operationId":"AuthController_resetPassword","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResetPasswordDto"}}}},"responses":{"200":{"description":"Reset password","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResetPasswordDto"}}}}},"tags":["auth"]}},"/auth/me":{"get":{"operationId":"AuthController_getAuthenticatedUser","parameters":[],"responses":{"200":{"description":"get authenticated user data","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuthResponseDto"}}}}},"tags":["auth"],"security":[{"bearer":[]}]}},"/generate-pdf":{"get":{"operationId":"GeneratePdfController_getPdf","parameters":[],"responses":{"200":{"description":""}},"tags":["generate-pdf"]}},"/product-details":{"post":{"operationId":"ProductDetailController_create","parameters":[],"responses":{"201":{"description":""}}},"get":{"operationId":"ProductDetailController_findAll","parameters":[],"responses":{"200":{"description":""}}}},"/product-details/{id}":{"get":{"operationId":"ProductDetailController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}}},"patch":{"operationId":"ProductDetailController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}}},"delete":{"operationId":"ProductDetailController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}}}},"/product-options":{"post":{"operationId":"ProductOptionController_create","parameters":[],"responses":{"201":{"description":""}}},"get":{"operationId":"ProductOptionController_findAll","parameters":[],"responses":{"200":{"description":""}}}},"/product-options/{id}":{"get":{"operationId":"ProductOptionController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}}},"patch":{"operationId":"ProductOptionController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}}},"delete":{"operationId":"ProductOptionController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}}}},"/product-collections":{"post":{"operationId":"ProductCollectionController_create","parameters":[],"responses":{"201":{"description":""}},"tags":["product-collections"],"security":[{"bearer":[]}]},"get":{"operationId":"ProductCollectionController_findAll","parameters":[],"responses":{"200":{"description":""}},"tags":["product-collections"],"security":[{"bearer":[]}]}},"/product-collections/search":{"get":{"operationId":"ProductCollectionController_search","parameters":[{"name":"collection_id","required":true,"in":"query","schema":{"type":"string"}},{"name":"search_keyword","required":true,"in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["product-collections"],"security":[{"bearer":[]}]}},"/product-collections/inspect":{"post":{"operationId":"ProductCollectionController_inspectItems","parameters":[],"responses":{"201":{"description":""}},"tags":["product-collections"],"security":[{"bearer":[]}]}},"/product-collections/inspect-delete":{"post":{"operationId":"ProductCollectionController_deleteItems","parameters":[],"responses":{"201":{"description":""}},"tags":["product-collections"],"security":[{"bearer":[]}]}},"/product-collections/{id}":{"get":{"operationId":"ProductCollectionController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["product-collections"],"security":[{"bearer":[]}]},"patch":{"operationId":"ProductCollectionController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["product-collections"],"security":[{"bearer":[]}]},"delete":{"operationId":"ProductCollectionController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["product-collections"],"security":[{"bearer":[]}]}},"/user-product-details":{"post":{"operationId":"UserProductDetailController_create","parameters":[],"responses":{"201":{"description":""}}},"get":{"operationId":"UserProductDetailController_findAll","parameters":[{"name":"page","required":true,"in":"query","schema":{"type":"string"}},{"name":"pageSize","required":true,"in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":""}}}},"/user-product-details/{id}":{"get":{"operationId":"UserProductDetailController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}}},"patch":{"operationId":"UserProductDetailController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}}},"delete":{"operationId":"UserProductDetailController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}}}},"/markets":{"post":{"operationId":"MarketController_create","parameters":[],"responses":{"201":{"description":""}}},"get":{"operationId":"MarketController_findAll","parameters":[{"name":"page","required":true,"in":"query","schema":{"type":"string"}},{"name":"pageSize","required":true,"in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":""}}}},"/markets/{id}":{"get":{"operationId":"MarketController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}}},"patch":{"operationId":"MarketController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}}},"delete":{"operationId":"MarketController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}}}},"/stores":{"post":{"operationId":"StoreController_create","parameters":[],"responses":{"201":{"description":""}}},"get":{"operationId":"StoreController_findAll","parameters":[{"name":"page","required":true,"in":"query","schema":{"type":"string"}},{"name":"pageSize","required":true,"in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":""}}}},"/stores/{id}":{"get":{"operationId":"StoreController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}}},"patch":{"operationId":"StoreController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}}},"delete":{"operationId":"StoreController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}}}},"/market-api-keys":{"post":{"operationId":"MarketApiKeyController_create","parameters":[],"responses":{"201":{"description":""}}},"get":{"operationId":"MarketApiKeyController_findAll","parameters":[{"name":"page","required":true,"in":"query","schema":{"type":"string"}},{"name":"pageSize","required":true,"in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":""}}}},"/market-api-keys/{id}":{"get":{"operationId":"MarketApiKeyController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}}},"patch":{"operationId":"MarketApiKeyController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}}},"delete":{"operationId":"MarketApiKeyController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}}}},"/translation-api-key":{"post":{"operationId":"TranslationApiKeyController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateTranslationApiKeyDto"}}}},"responses":{"201":{"description":""}},"tags":["Translation API Key"]},"get":{"operationId":"TranslationApiKeyController_findAll","parameters":[],"responses":{"200":{"description":""}},"tags":["Translation API Key"]}},"/translation-api-key/check":{"post":{"operationId":"TranslationApiKeyController_check","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateTranslationApiKeyDto"}}}},"responses":{"200":{"description":"users response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResponseDto"}}}}},"tags":["Translation API Key"]}},"/translation-api-key/{id}":{"get":{"operationId":"TranslationApiKeyController_findBySiteId","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Translation API Key"]},"put":{"operationId":"TranslationApiKeyController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateTranslationApiKeyDto"}}}},"responses":{"200":{"description":""}},"tags":["Translation API Key"]},"delete":{"operationId":"TranslationApiKeyController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Translation API Key"]}},"/posts":{"get":{"operationId":"PostsController_getPosts","parameters":[{"name":"boardId","required":true,"in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":""}}},"post":{"operationId":"PostsController_createPost","parameters":[],"responses":{"201":{"description":""}}}},"/posts/{id}":{"get":{"operationId":"PostsController_getPost","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}}},"put":{"operationId":"PostsController_updatePost","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}}},"delete":{"operationId":"PostsController_deletePost","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}}}},"/board/list":{"get":{"operationId":"BoardController_findAll","parameters":[],"responses":{"200":{"description":""}},"tags":["게시판"]}},"/board/{id}":{"get":{"operationId":"BoardController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["게시판"]}},"/sourcing-sites":{"post":{"operationId":"SourcingSiteController_create","parameters":[],"responses":{"201":{"description":""}},"tags":["sourcing-site"],"security":[{"bearer":[]}]},"get":{"operationId":"SourcingSiteController_findAll","parameters":[{"name":"page","required":true,"in":"query","schema":{"type":"string"}},{"name":"pageSize","required":true,"in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["sourcing-site"],"security":[{"bearer":[]}]}},"/sourcing-sites/sellian":{"get":{"operationId":"SourcingSiteController_findAllSellian","parameters":[],"responses":{"200":{"description":""}},"tags":["sourcing-site"],"security":[{"bearer":[]}]}},"/sourcing-sites/crosell":{"get":{"operationId":"SourcingSiteController_findAllCrosell","parameters":[],"responses":{"200":{"description":""}},"tags":["sourcing-site"],"security":[{"bearer":[]}]}},"/sourcing-sites/{id}":{"get":{"operationId":"SourcingSiteController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["sourcing-site"],"security":[{"bearer":[]}]},"patch":{"operationId":"SourcingSiteController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["sourcing-site"],"security":[{"bearer":[]}]},"delete":{"operationId":"SourcingSiteController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["sourcing-site"],"security":[{"bearer":[]}]}},"/sourcing/request-list":{"post":{"operationId":"SourcingController_RequestList","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ReqListDto"}}}},"responses":{"201":{"description":""}},"tags":["sourcing"],"security":[{"bearer":[]}]},"get":{"operationId":"SourcingController_GetRequestList","parameters":[],"responses":{"200":{"description":""}},"tags":["sourcing"],"security":[{"bearer":[]}]}},"/sourcing/request-detail":{"post":{"operationId":"SourcingController_RequestDetail","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RequestDetailDto"}}}},"responses":{"201":{"description":""}},"tags":["sourcing"],"security":[{"bearer":[]}]}}},"info":{"title":"LOG","description":"crawler server apis","version":"1.0","contact":{}},"tags":[],"servers":[],"components":{"securitySchemes":{"bearer":{"scheme":"bearer","bearerFormat":"JWT","type":"http"}},"schemas":{"ResponseDto":{"type":"object","properties":{"data":{"type":"object"}},"required":["data"]},"RequestCollectionListDto":{"type":"object","properties":{"start_date":{"type":"string"},"end_date":{"type":"string"}},"required":["start_date","end_date"]},"EmailDTO":{"type":"object","properties":{"name":{"type":"string"},"to":{"type":"string"},"subject":{"type":"string"},"verify":{"type":"string"}},"required":["name","to","subject"]},"UserDTO":{"type":"object","properties":{"name":{"type":"string"},"phone":{"type":"string"},"email":{"type":"string"},"password":{"type":"string"},"businessNumber":{"type":"string"},"businessLicense":{"type":"string"}},"required":["name","phone","email","password","businessNumber","businessLicense"]},"LoginDto":{"type":"object","properties":{"email":{"type":"string"},"password":{"type":"string"}},"required":["email","password"]},"AuthResponseDto":{"type":"object","properties":{"accessToken":{"type":"object"},"refreshToken":{"type":"object"},"user":{"type":"object"}},"required":["accessToken","refreshToken","user"]},"RefreshTokenDto":{"type":"object","properties":{"refreshToken":{"type":"string"}},"required":["refreshToken"]},"ResetPasswordReqDto":{"type":"object","properties":{"email":{"type":"string"}},"required":["email"]},"ResetPasswordDto":{"type":"object","properties":{"newPassword":{"type":"string"},"resetPasswordToken":{"type":"string"}},"required":["newPassword","resetPasswordToken"]},"CreateTranslationApiKeyDto":{"type":"object","properties":{"translationSiteId":{"type":"number","description":"번역 사이트 ID"},"apiKey":{"type":"string","description":"번역 API Key"}},"required":["translationSiteId","apiKey"]},"UpdateTranslationApiKeyDto":{"type":"object","properties":{"translationSiteId":{"type":"number","description":"번역 사이트 ID"},"apiKey":{"type":"string","description":"번역 API Key"}}},"PageScopeDto":{"type":"object","properties":{"start_page":{"type":"number","default":1},"end_page":{"type":"number","default":1}},"required":["start_page","end_page"]},"ReqOptionDto":{"type":"object","properties":{"allowAdult":{"type":"boolean"},"allowAd":{"type":"boolean"},"allowOriginalPrice":{"type":"boolean"},"allowBanword":{"type":"boolean"},"allowBancode":{"type":"boolean"},"limitCount":{"type":"number"},"marginTemplateId":{"type":"string"},"categoryTemplateId":{"type":"string"},"tagTemplateId":{"type":"string"},"forbiddenWordTemplateId":{"type":"string"},"replaceWordTemplateId":{"type":"string"},"translate":{"type":"string","default":"google"}},"required":["allowAdult","allowAd","allowOriginalPrice","allowBanword","allowBancode","limitCount","marginTemplateId","categoryTemplateId","tagTemplateId","forbiddenWordTemplateId","replaceWordTemplateId","translate"]},"ReqListDto":{"type":"object","properties":{"userId":{"type":"string"},"crawler_name":{"type":"string","default":"crawler_1"},"target":{"type":"string","default":"amazon"},"type":{"type":"string"},"search_keyword":{"type":"string"},"request_url":{"type":"string","default":"https://www.amazon.com/s?k=nike+dunks+women&crid=2H9OU7VJ21SF4&sprefix=nike+d%2Caps%2C314&ref=nb_sb_ss_pltr-mrr_3_6"},"proxy_mode":{"type":"string","default":"true"},"page_scope":{"$ref":"#/components/schemas/PageScopeDto"},"option":{"$ref":"#/components/schemas/ReqOptionDto"},"sourcingSiteId":{"type":"number"}},"required":["userId","crawler_name","target","type","search_keyword","request_url","proxy_mode","page_scope","option","sourcingSiteId"]},"RequestDetailDto":{"type":"object","properties":{"collectionId":{"type":"string","description":"collection ID(수집그룹아이디)"},"userId":{"type":"string","description":"무시"},"productId":{"description":"상품 id( In table id)","type":"array","items":{"type":"string"}}},"required":["collectionId","userId","productId"]}}}}